+++ crates/matrix-sdk-crypto/src/store/mod.rs
@@ -52,8 +52,8 @@ use futures_core::Stream;
 use futures_util::StreamExt;
 use itertools::{Either, Itertools};
 use ruma::{
-    DeviceId, OwnedDeviceId, OwnedUserId, RoomId, UserId, encryption::KeyUsage,
-    events::secret::request::SecretName,
+    DeviceId, MilliSecondsSinceUnixEpoch, OwnedDeviceId, OwnedUserId, RoomId, UserId,
+    encryption::KeyUsage, events::secret::request::SecretName,
 };
 use serde::{Serialize, de::DeserializeOwned};
 use thiserror::Error;
@@ -65,7 +65,8 @@ use vodozemac::{Curve25519PublicKey, megolm::SessionOrdering};
 
 use self::types::{
     Changes, CrossSigningKeyExport, DeviceChanges, DeviceUpdates, IdentityChanges, IdentityUpdates,
-    PendingChanges, RoomKeyInfo, RoomKeyWithheldInfo, UserKeyQueryResult,
+    PendingChanges, RoomKeyInfo, RoomKeyWithheldInfo, RoomPendingKeyBundleDetails,
+    UserKeyQueryResult,
 };
 use crate::{
     CrossSigningStatus, OwnUserIdentityData, RoomKeyImportResult,
@@ -1839,6 +1840,44 @@ impl Store {
         );
 
         Ok(())
+    }
+
+    /// Store the fact that we have accepted an invite for a given room on this
+    /// client, so should accept an [MSC4268] key bundle if one arrives
+    /// soon.
+    ///
+    /// [MSC4268]: https://github.com/matrix-org/matrix-spec-proposals/pull/4268
+    pub async fn store_room_pending_key_bundle(
+        &self,
+        room_id: &RoomId,
+        inviter: &UserId,
+    ) -> Result<(), CryptoStoreError> {
+        let invite_accepted_at = MilliSecondsSinceUnixEpoch::now();
+        self.save_changes(Changes {
+            rooms_pending_key_bundle: HashMap::from([(
+                room_id.to_owned(),
+                Some(RoomPendingKeyBundleDetails {
+                    room_id: room_id.to_owned(),
+                    invite_accepted_at,
+                    inviter: inviter.to_owned(),
+                }),
+            )]),
+            ..Default::default()
+        })
+        .await?;
+
+        Ok(())
+    }
+
+    /// Clear the record of accepting an invite for a specific room.
+    ///
+    /// The counterpart of [`Store::store_room_pending_key_bundle`].
+    pub async fn clear_room_pending_key_bundle(&self, room_id: &RoomId) -> Result<()> {
+        self.save_changes(Changes {
+            rooms_pending_key_bundle: HashMap::from([(room_id.to_owned(), None)]),
+            ..Default::default()
+        })
+        .await
     }
 }