+++ crates/matrix-sdk-indexeddb/src/crypto_store/migrations/mod.rs
@@ -31,6 +31,7 @@ mod old_keys;
 mod v0_to_v5;
 mod v101_to_v102;
 mod v102_to_v103;
+mod v103_to_v104;
 mod v10_to_v11;
 mod v11_to_v12;
 mod v12_to_v13;
@@ -196,6 +197,10 @@ pub async fn open_and_upgrade_db(
         v102_to_v103::schema_add(name).await?;
     }
 
+    if old_version < 104 {
+        v103_to_v104::schema_add(name).await?;
+    }
+
     // If you add more migrations here, you'll need to update
     // `tests::EXPECTED_SCHEMA_VERSION`.
 
@@ -300,7 +305,7 @@ mod tests {
     wasm_bindgen_test::wasm_bindgen_test_configure!(run_in_browser);
 
     /// The schema version we expect after we open the store.
-    const EXPECTED_SCHEMA_VERSION: u32 = 103;
+    const EXPECTED_SCHEMA_VERSION: u32 = 104;
 
     /// Adjust this to test do a more comprehensive perf test
     const NUM_RECORDS_FOR_PERF: usize = 2_000;