+++ crates/matrix-sdk-indexeddb/src/crypto_store/migrations/v103_to_v104.rs
@@ -12,21 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use eyeball_im::VectorDiff;
-use matrix_sdk_base::event_cache::Event;
+use indexed_db_futures::{Build, error::OpenDbError};
 
-use crate::event_cache::EventsOrigin;
+use crate::crypto_store::{Result, keys, migrations::do_schema_upgrade};
 
-pub(super) mod lock;
-pub mod pagination;
-pub mod room;
-pub mod thread;
-
-/// A diff update for an event cache timeline represented as a vector.
-#[derive(Clone, Debug)]
-pub struct TimelineVectorDiffs {
-    /// New vector diff for the thread timeline.
-    pub diffs: Vec<VectorDiff<Event>>,
-    /// The origin that triggered this update.
-    pub origin: EventsOrigin,
+/// Perform the schema upgrade v013 to v104: add the
+/// `rooms_pending_key_bundle` store.
+pub(crate) async fn schema_add(name: &str) -> Result<(), OpenDbError> {
+    do_schema_upgrade(name, 104, |tx, _| {
+        tx.db().create_object_store(keys::ROOMS_PENDING_KEY_BUNDLE).build()?;
+        Ok(())
+    })
+    .await
 }