+++ crates/matrix-sdk/src/event_cache/mod.rs
@@ -73,14 +73,16 @@ use crate::{
 
 mod caches;
 mod deduplicator;
-mod pagination;
 mod persistence;
 #[cfg(feature = "e2e-encryption")]
 mod redecryptor;
 mod room;
 
-pub use caches::TimelineVectorDiffs;
-pub use pagination::{RoomPagination, RoomPaginationStatus};
+pub use caches::{
+    TimelineVectorDiffs,
+    pagination::{BackPaginationOutcome, PaginationStatus},
+    room::pagination::RoomPagination,
+};
 #[cfg(feature = "e2e-encryption")]
 pub use redecryptor::{DecryptionRetryRequest, RedecryptorReport};
 pub use room::{RoomEventCache, RoomEventCacheSubscriber};
@@ -1101,7 +1103,7 @@ impl EventCacheInner {
                 }
 
                 let pagination_status =
-                    SharedObservable::new(RoomPaginationStatus::Idle { hit_timeline_start: false });
+                    SharedObservable::new(PaginationStatus::Idle { hit_timeline_start: false });
 
                 let Some(client) = self.client.get() else {
                     return Err(EventCacheError::ClientDropped);
@@ -1168,21 +1170,6 @@ impl EventCacheInner {
             }
         }
     }
-}
-
-/// The result of a single back-pagination request.
-#[derive(Debug)]
-pub struct BackPaginationOutcome {
-    /// Did the back-pagination reach the start of the timeline?
-    pub reached_start: bool,
-
-    /// All the events that have been returned in the back-pagination
-    /// request.
-    ///
-    /// Events are presented in reverse order: the first element of the vec,
-    /// if present, is the most "recent" event from the chunk (or
-    /// technically, the last one in the topological ordering).
-    pub events: Vec<TimelineEvent>,
 }
 
 /// Represents a timeline update of a room. It hides the details of