+++ crates/matrix-sdk/src/event_cache/room/threads.rs
@@ -24,9 +24,9 @@ use ruma::{EventId, OwnedEventId, OwnedRoomId};
 use tokio::sync::broadcast::{Receiver, Sender};
 use tracing::{error, trace};
 
-use crate::event_cache::{
-    BackPaginationOutcome, EventsOrigin, RoomEventCacheLinkedChunkUpdate,
-    caches::TimelineVectorDiffs,
+use super::super::{
+    EventsOrigin, RoomEventCacheLinkedChunkUpdate,
+    caches::{TimelineVectorDiffs, pagination::BackPaginationOutcome},
     deduplicator::DeduplicationOutcome,
     room::{LoadMoreEventsBackwardsOutcome, events::EventLinkedChunk},
 };
@@ -173,7 +173,12 @@ impl ThreadEventCache {
         // resolve the gap.
         if let Some(prev_token) = self.chunk.rgap().map(|gap| gap.prev_token) {
             trace!(%prev_token, "thread chunk has at least a gap");
-            return LoadMoreEventsBackwardsOutcome::Gap { prev_token: Some(prev_token) };
+            return LoadMoreEventsBackwardsOutcome::Gap {
+                prev_token: Some(prev_token),
+                // Since there is `Some(prev_token)` already, we assume we've
+                // waited for it already.
+                waited_for_initial_prev_token: true,
+            };
         }
 
         // If we don't have a gap, then the first event should be the the thread's root;
@@ -190,7 +195,11 @@ impl ThreadEventCache {
 
         // Otherwise, we don't have a gap nor events. We don't have anything. Poor us.
         // Well, is ok: start a pagination from the end.
-        LoadMoreEventsBackwardsOutcome::Gap { prev_token: None }
+        LoadMoreEventsBackwardsOutcome::Gap {
+            prev_token: None,
+            // No `prev_token` for threads, let's assume it's been waited.
+            waited_for_initial_prev_token: true,
+        }
     }
 
     /// Find duplicates in a thread, until there's persistent storage for